requirements.txt
Django==2.0.4
djangorestframework==3.8.2
djangorestframework-jwt==1.11.0
django-split-settings==0.3.0
pylint==1.6.5
pylint==1.8.4
